requirements.txt
torch>=2.1.0
torchvision>=0.16.0
albumentations>=1.4.0
mlflow>=2.9.0
gradio>=4.7.0
opencv-python-headless>=4.8.0
scikit-image>=0.21.0
lime>=0.2.0.1
PyYAML>=6.0.1
tqdm>=4.65.0
numpy>=1.24.0
pillow>=10.0.0
matplotlib>=3.7.0
pandas>=2.1.0
huggingface-hub>=0.20.0
